# packages/serve/openapi.json
{"openapi":"3.1.0","info":{"title":"@abtree/serve","description":"Development documentation","version":"0.1.0"},"paths":{"/api/executions":{"get":{"operationId":"getExecutions","summary":"List executions","tags":["Executions"],"responses":{"200":{"description":"Execution summaries","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ExecutionSummary"}}}}}}}},"/api/executions/{id}":{"get":{"operationId":"getExecution","summary":"Get a single execution","tags":["Executions"],"parameters":[{"in":"path","name":"id","schema":{"type":"string"},"required":true}],"responses":{"200":{"description":"Execution doc","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Execution"}}}},"404":{"description":"Unknown execution id","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorData"}}}}}}},"/api/executions/{id}/tree":{"get":{"operationId":"getExecutionTree","summary":"ReactFlow graph for an execution's tree snapshot","tags":["Executions"],"parameters":[{"in":"path","name":"id","schema":{"type":"string"},"required":true}],"responses":{"200":{"description":"Tree graph","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TreeGraph"}}}},"404":{"description":"Unknown execution id or missing snapshot","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorData"}}}}}}}},"components":{"schemas":{"ExecutionSummary":{"type":"object","properties":{"id":{"type":"string"},"tree":{"type":"string"},"summary":{"type":"string"},"status":{"type":"string"},"phase":{"type":"string"},"cursor":{"type":"string"},"snapshot":{"type":"string"},"protocolAccepted":{"type":"boolean"},"traceCount":{"type":"integer","minimum":0,"maximum":9007199254740991},"createdAt":{"type":"string"},"updatedAt":{"type":"string"}},"required":["id","tree","summary","status","phase","cursor","snapshot","protocolAccepted","traceCount","createdAt","updatedAt"]},"TraceEntry":{"type":"object","properties":{"ts":{"type":"string"},"kind":{"type":"string","enum":["evaluate","instruct","protocol"]},"cursor":{"type":"string"},"name":{"type":"string"},"submitted":{"type":"string"},"outcome":{"type":"string"},"note":{"type":"string"}},"required":["ts","kind","cursor","name","submitted","outcome"]},"Execution":{"type":"object","properties":{"id":{"type":"string"},"tree":{"type":"string"},"summary":{"type":"string"},"status":{"type":"string"},"phase":{"type":"string"},"cursor":{"type":"string"},"snapshot":{"type":"string"},"protocolAccepted":{"type":"boolean"},"trace":{"type":"array","items":{"$ref":"#/components/schemas/TraceEntry"}},"local":{"type":"object","propertyNames":{"type":"string"},"additionalProperties":{}},"global":{"type":"object","propertyNames":{"type":"string"},"additionalProperties":{}},"runtime":{"type":"object","properties":{"nodeStatus":{"type":"object","propertyNames":{"type":"string"},"additionalProperties":{"type":"string","enum":["success","failure","running"]}},"stepIndex":{"type":"object","propertyNames":{"type":"string"},"additionalProperties":{"type":"integer","minimum":-9007199254740991,"maximum":9007199254740991}},"retryCount":{"type":"object","propertyNames":{"type":"string"},"additionalProperties":{"type":"integer","minimum":-9007199254740991,"maximum":9007199254740991}}},"required":["nodeStatus","stepIndex","retryCount"]},"createdAt":{"type":"string"},"updatedAt":{"type":"string"}},"required":["id","tree","summary","status","phase","cursor","snapshot","protocolAccepted","trace","local","global","runtime","createdAt","updatedAt"]},"ErrorData":{"type":"object","properties":{"timestamp":{"type":"string"},"status":{"type":"number","enum":[100,101,102,103,200,201,202,203,204,205,206,207,300,301,302,303,304,305,307,308,400,401,402,403,404,405,406,407,408,409,410,411,412,413,414,415,416,417,418,419,420,421,422,423,424,426,428,429,431,451,500,501,502,503,504,505,507,511]},"error":{"type":"string"},"message":{"type":"string"}},"required":["timestamp","status","error","message"]},"TreeNodeData":{"type":"object","properties":{"id":{"type":"string"},"path":{"type":"string"},"name":{"type":"string"},"kind":{"type":"string","enum":["action","sequence","selector","parallel","ref"]},"ref":{"anyOf":[{"type":"string"},{"type":"null"}]},"steps":{"type":"array","items":{"type":"object","properties":{"kind":{"type":"string","enum":["evaluate","instruct"]},"text":{"type":"string"}},"required":["kind","text"]}},"status":{"anyOf":[{"type":"string","enum":["success","failure","running"]},{"type":"null"}]},"retries":{"type":"integer","minimum":-9007199254740991,"maximum":9007199254740991},"x":{"type":"number"},"y":{"type":"number"}},"required":["id","path","name","kind","ref","status","x","y"]},"TreeEdgeData":{"type":"object","properties":{"id":{"type":"string"},"source":{"type":"string"},"target":{"type":"string"}},"required":["id","source","target"]},"TreeGraph":{"type":"object","properties":{"nodes":{"type":"array","items":{"$ref":"#/components/schemas/TreeNodeData"}},"edges":{"type":"array","items":{"$ref":"#/components/schemas/TreeEdgeData"}},"width":{"type":"number"},"height":{"type":"number"}},"required":["nodes","edges","width","height"]}}}}